// server/src/application/dtos/plan/request/create-plan.dto.ts
export interface CreatePlanRequestDto {
    name: string;
    price: number;
    currency?: string;
    durationDays: number;
    maxWorkspaces: number;
    maxMembersPerWorkspace: number;
    messageRetentionDays: number;
    aiAssistantEnabled?: boolean;
    videoCallsEnabled?: boolean;
    multiAiAgents?: boolean;
    pinBoardEnabled?: boolean;
    isActive?: boolean;
}